move md5 logic to datasetfile
